Successfully added a new ui to download plugins.
UI is similar to the existing plugins page. It'll have to renamed to
'Installed Plugins' or something similar. The two could also be grouped
together under a heading 'Plugins' -> 'Installed' , 'Download'.

Some of the data in QTreeWidget is loaded via a JSON. While the other is
being randomly generated for now.

The JSON will be available to Picard via an API or as a static file.

The filter plugins option can be used to search for a specific plugin,
I'm hoping that this can support regex.

The checkboxes can be used to download multiple plugins at once.

TODO:

I need to figure out how to update the Details field now, And the
prototype will be mostly done.

The focus would then move on to the code-behind.